refactor(low-code): trim config_overrides docs and move its tests
No behaviour change. Three review points on style.

The docstrings had grown into a transcript of the review that produced them - rationale for decisions
already made, arguments against alternatives nobody will propose again. That rots in a shared file.
Trimmed to the semantics a caller needs, with the one open caveat pointed at its issue rather than
restated: airbytehq/airbyte-internal-issues#16995. Docstring-to-logic ratio across the eight new
methods is now 68:109, down from roughly 200:120.

`_CONFIG_PERSISTING_AUTHENTICATOR_TYPES` sat between two methods; moved up beside
`_LOWEST_SAFE_CONCURRENCY_LEVEL` with the other class-level constants. `_FIELDS_HOLDING_CONFIG_VALUES`
stays in the resolver, since that is the module that consumes it.

Moved the 30 tests out of `checks/test_check_stream.py`, which was the wrong home: the behaviour is in
`ConcurrentDeclarativeSource` and the check factories deliberately ignore `model.config_overrides`.
They live in `test_concurrent_declarative_source_config_overrides.py`, beside
`test_concurrent_declarative_source.py` rather than inside it, because that module is already 6.1k
lines and this would have pushed it past 7k.

Re-ran four mutations after the move to confirm the tests still bind to the code and not to their old
location: dropping `CustomAuthenticator` from the type set, reverting the type test to `isinstance`,
removing the resolver exemption, and removing the secret registration each fail exactly the tests that
name them.
